+++ contracts/Ticket/Verifier_Producer.sol
@@ -12,7 +12,8 @@ import "../Fund/FundManager.sol";
 import "./LuckManager.sol";
 import "@openzeppelin/upgrades/contracts/Initializable.sol";
 
-contract VerifierProducer is Initializable{
+
+contract VerifierProducer is Initializable {
     using SafeMath for uint256;
 
     Producer producerRegistry;
@@ -25,15 +26,15 @@ contract VerifierProducer is Initializable{
 
     mapping(bytes32 => bool) rewardedBlocks;
 
-    function initialize(address _producerRegistry, 
-                address _clusterRegistry, 
-                address _luckManager, 
-                address _pot, 
-                address _fundManager, 
-                bytes32 _producerRole,
-                bytes32 _tokenId) 
-                public 
-                initializer {
+    function initialize(
+        address _producerRegistry,
+        address _clusterRegistry,
+        address _luckManager,
+        address _pot,
+        address _fundManager,
+        bytes32 _producerRole,
+        bytes32 _tokenId
+    ) public initializer {
         producerRegistry = Producer(_producerRegistry);
         clusterRegistry = ClusterRegistry(_clusterRegistry);
         luckManager = LuckManager(_luckManager);
@@ -43,89 +44,143 @@ contract VerifierProducer is Initializable{
         tokenId = _tokenId;
     }
 
-    function verifyClaim(bytes memory _blockHeader, 
-                        bytes memory _relayerSig, 
-                        bytes memory _producerSig,
-                        address _cluster,
-                        bool _isAggregated) 
-                        public 
-                        returns(bytes32, address, uint) {
+    function verifyClaim(
+        bytes memory _blockHeader,
+        bytes memory _relayerSig,
+        bytes memory _producerSig,
+        address _cluster,
+        bool _isAggregated
+    )
+        public
+        returns (
+            bytes32,
+            address,
+            uint256
+        )
+    {
         bytes32 blockHash = keccak256(_blockHeader);
         require(!rewardedBlocks[blockHash], "Block header already rewarded");
         bytes memory coinBase = extractCoinBase(_blockHeader);
-        uint blockNumber = extractBlockNumber(_blockHeader);
+        uint256 blockNumber = extractBlockNumber(_blockHeader);
         address actualProducer = producerRegistry.getProducer(coinBase);
         address relayer = recoverSigner(blockHash, _relayerSig);
 
-        require(uint(clusterRegistry.getClusterStatus(_cluster)) == 2, "Verifier_Producer: Cluster isn't active");
-        require(Cluster(_cluster).isRelayer(relayer), "Verifier_Producer: Relayer isn't part of cluster");
+        require(
+            uint256(clusterRegistry.getClusterStatus(_cluster)) == 2,
+            "Verifier_Producer: Cluster isn't active"
+        );
+        require(
+            Cluster(_cluster).isRelayer(relayer),
+            "Verifier_Producer: Relayer isn't part of cluster"
+        );
 
-        bytes memory producerSigPayload = abi.encodePacked(_blockHeader, _relayerSig);
-        address extractedProducer = recoverSigner(keccak256(producerSigPayload), _producerSig);
-        require(extractedProducer == actualProducer, "Verifier_Producer: Producer sig doesn't match coinbase");
-        bytes32 ticket = keccak256(abi.encodePacked(producerSigPayload, _producerSig));
-        uint epoch = pot.getEpoch(blockNumber);
-        uint luckLimit = luckManager.getLuck(epoch, producerRole);
-        require(uint(luckLimit) > uint(ticket), "Verifier_Producer: Ticket not in winning range");
-        if(pot.getPotValue(epoch, tokenId) == 0) {
-            uint[] memory epochs;
-            uint[] memory values;
-            uint[][] memory inflationLog = fundManager.draw(address(pot), block.number);
-            for(uint i=0; i < inflationLog[0].length.sub(1); i++) {
-                for(uint j=inflationLog[0][i]; j < inflationLog[0][i+1]; j++) {
+        bytes memory producerSigPayload = abi.encodePacked(
+            _blockHeader,
+            _relayerSig
+        );
+        address extractedProducer = recoverSigner(
+            keccak256(producerSigPayload),
+            _producerSig
+        );
+        require(
+            extractedProducer == actualProducer,
+            "Verifier_Producer: Producer sig doesn't match coinbase"
+        );
+        bytes32 ticket = keccak256(
+            abi.encodePacked(producerSigPayload, _producerSig)
+        );
+        uint256 epoch = pot.getEpoch(blockNumber);
+        uint256 luckLimit = luckManager.getLuck(epoch, producerRole);
+        require(
+            uint256(luckLimit) > uint256(ticket),
+            "Verifier_Producer: Ticket not in winning range"
+        );
+        if (pot.getPotValue(epoch, tokenId) == 0) {
+            uint256[] memory epochs;
+            uint256[] memory values;
+            uint256[][] memory inflationLog = fundManager.draw(
+                address(pot),
+                block.number
+            );
+            for (uint256 i = 0; i < inflationLog[0].length.sub(1); i++) {
+                for (
+                    uint256 j = inflationLog[0][i];
+                    j < inflationLog[0][i + 1];
+                    j++
+                ) {
                     epochs[epochs.length] = j;
                     values[values.length] = inflationLog[1][i];
                 }
             }
-            require(pot.addToPot(epochs, address(fundManager), tokenId, values), 
-                "Verifier_Producer: Could not add to pot");
+            require(
+                pot.addToPot(epochs, address(fundManager), tokenId, values),
+                "Verifier_Producer: Could not add to pot"
+            );
         }
         //TODO: If encoderv2 can be used then remove isAggregated
-        if(!_isAggregated) {
+        if (!_isAggregated) {
             //TODO: Find if there is a better way to initialize dynamic arrays
             bytes32[] memory roles;
             address[] memory claimers;
-            uint[] memory epochs;
+            uint256[] memory epochs;
             roles[0] = producerRole;
             claimers[0] = actualProducer;
             epochs[0] = epoch;
-            require(pot.claimTicket(roles, claimers, epochs), 
-                    "Verifier_Producer: Ticket claim failed");
+            require(
+                pot.claimTicket(roles, claimers, epochs),
+                "Verifier_Producer: Ticket claim failed"
+            );
         }
         rewardedBlocks[blockHash] = true;
         return (producerRole, actualProducer, epoch);
     }
 
-    function verifyClaims(bytes[] memory _blockHeaders, 
-                          bytes[] memory _relayerSigs, 
-                          bytes[] memory _producerSigs, 
-                          address[] memory _clusters) 
-                          public {
-        require(_blockHeaders.length == _relayerSigs.length && 
+    function verifyClaims(
+        bytes[] memory _blockHeaders,
+        bytes[] memory _relayerSigs,
+        bytes[] memory _producerSigs,
+        address[] memory _clusters
+    ) public {
+        require(
+            _blockHeaders.length == _relayerSigs.length &&
                 _relayerSigs.length == _producerSigs.length &&
-                _producerSigs.length == _clusters.length
-                , "Verifier_Producer: Invalid Inputs");
+                _producerSigs.length == _clusters.length,
+            "Verifier_Producer: Invalid Inputs"
+        );
         bytes32[] memory roles;
         address[] memory claimers;
-        uint[] memory epochs;
-        for(uint i=0; i < _blockHeaders.length; i++) {
-            (bytes32 role, address claimer, uint epoch) = verifyClaim(_blockHeaders[i], 
-                                                                        _relayerSigs[i], 
-                                                                        _producerSigs[i],
-                                                                        _clusters[i],
-                                                                        true);
+        uint256[] memory epochs;
+        for (uint256 i = 0; i < _blockHeaders.length; i++) {
+            (bytes32 role, address claimer, uint256 epoch) = verifyClaim(
+                _blockHeaders[i],
+                _relayerSigs[i],
+                _producerSigs[i],
+                _clusters[i],
+                true
+            );
             epochs[epochs.length] = epoch;
             claimers[claimers.length] = claimer;
             roles[roles.length] = role;
         }
-        require(pot.claimTicket(roles, claimers, epochs), "Verifier_Producer: Aggregate ticket claim failed");
+        require(
+            pot.claimTicket(roles, claimers, epochs),
+            "Verifier_Producer: Aggregate ticket claim failed"
+        );
     }
 
-    function extractCoinBase(bytes memory blockHeader) public view returns(bytes memory) {
+    function extractCoinBase(bytes memory blockHeader)
+        public
+        view
+        returns (bytes memory)
+    {
         // TODO: Implementation specific for blockchain
     }
 
-    function extractBlockNumber(bytes memory blockHeader) public view returns(uint) {
+    function extractBlockNumber(bytes memory blockHeader)
+        public
+        view
+        returns (uint256)
+    {
         // TODO: Implementation specific for blockchain
     }
 
@@ -133,7 +188,11 @@ contract VerifierProducer is Initializable{
     function splitSignature(bytes memory sig)
         internal
         pure
-        returns (uint8 v, bytes32 r, bytes32 s)
+        returns (
+            uint8 v,
+            bytes32 r,
+            bytes32 s
+        )
     {
         require(sig.length == 65);
